Finish the start screen.
I wanted to add two demo models of the different character types. However, as part of that I realized I need to change the mouse cursor and that I had multiple sprite animations in a file. As a result, I had to split
out the separate animations to use them correctly.